Improve the HttpConnector class with an executor
We have decided to use an executor to pass requests to the
HttpProcessor because we don't want the connector thread to ever be
blocked. Before, it could happen in the below scenario if this class
was changed to use only one HttpProcessor object:
1. The connector thread assigns a socket to the processor thread.
2. The run method of the HttpProcessor object receives the socket via
its await method. But after the socket is received, the run method of
the processor thread takes long to process the request. In the mean
time, it is suspended by the OS.
3. A second socket is assigned to the HttpProcessor object. The
available flag of the later object is set to true.
4. A third request is assigned to the HttpProcessor object. But because
the available flag was already set to true, the wait method is called
on the HttpProcessor object. This suspends the connector thread which
will have to wait for the processor thread to be resumed and to call
the notifyAll method once it has finished processing the first request,
and started to process the second one.